pyproject.toml
[project]
name = "signal-radar"
version = "0.1.0"
description = "Swing trading signal engine for stocks and forex (daily trend following)"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0",
    "pandas>=2.2",
    "yfinance>=0.2.40",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
analysis = [
    "matplotlib>=3.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
